Rename + separate concerns
